+++ server/src/handlers/plugin_run.rs
@@ -13,7 +13,7 @@ use atomic_lib::{hierarchy::check_write, Storelike};
 use crate::{
     appstate::AppState,
     errors::{AtomicServerError, AtomicServerResult},
-    helpers::get_client_agent,
+    helpers::get_client_agent_of,
     plugins::js_runtime,
 };
 
@@ -62,7 +62,7 @@ pub async fn handle_plugin_run(
     let signed_subject =
         atomic_lib::Subject::from_raw(&path_and_query, None).resolve(&context.origin);
 
-    let agent = get_client_agent(req.headers(), &appstate, &signed_subject).await?;
+    let agent = get_client_agent_of(&req, &appstate, &signed_subject).await?;
     check_write(store, &resource, &agent).await?;
 
     let runtime = js_runtime::embedded_runtime()?;